rename: backend 디렉토리를 app 폴더 밖으로 이동
