+++ (backend)/auths/domains/entities/CommonAuthEntity.ts
