+++ (backend)/juniors/auths/domains/entities/JuniorAuth.ts
