+++ (backend)/seniors/auths/domains/entities/SeniorAuth.ts
